requirements.txt
pika==1.2.0
requests==2.26.0
beautifulsoup4==4.10.0
mysql-connector-python==8.0.28
pandas==1.3.3
